Add confirmation to delete reservation and restructured code for validation.

+++ restaurant/views.py
@@ -1,4 +1,9 @@
-from django.shortcuts import render, HttpResponseRedirect, get_object_or_404, redirect
+from django.shortcuts import (
+                                render,
+                                HttpResponseRedirect,
+                                get_object_or_404,
+                                redirect
+                                )
 from .models import Table, Reservation, Review, Food, Drink
 from django.contrib import messages
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
@@ -6,7 +11,7 @@ from django.contrib.auth.decorators import login_required
 from .forms import ReservationForm, ReviewForm
 from account.models import Profile
 from django.utils.timezone import now
-import datetime
+from datetime import datetime
 import pytz
 
 
@@ -39,31 +44,61 @@ def table_detail(request, id, slug):
     if request.method == 'POST':
         form = ReservationForm(request.POST)
         customer = request.user
-        reserve_start = request.POST['reserve_start']
-        reserve_end = request.POST['reserve_end']
-        if reserve_start == '' or reserve_end == '':
-            messages.error(request, 'Enter your desires times for reservation.')
+        r_start = request.POST['reserve_start']
+        r_end = request.POST['reserve_end']
+        if r_start == '' or r_end == '':
+            messages.error(
+                request,
+                'Enter your desires times for reservation.'
+                )
             return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
-        reserve_start = datetime.datetime.strptime(reserve_start, '%m/%d/%Y %I:%M %p').replace(tzinfo=pytz.utc)
-        reserve_end = datetime.datetime.strptime(reserve_end, '%m/%d/%Y %I:%M %p').replace(tzinfo=pytz.utc)
-        case_1 = Reservation.objects.filter(table=table, reserve_start__lte=reserve_start, reserve_end__gte=reserve_start).exists()
-        case_2 = Reservation.objects.filter(table=table, reserve_start__lte=reserve_end, reserve_end__gte=reserve_end).exists()
-        case_3 = Reservation.objects.filter(table=table, reserve_start__gte=reserve_start, reserve_end__lte=reserve_end).exists()
+        r_start = datetime.strptime(
+                                        r_start,
+                                        '%m/%d/%Y %I:%M %p'
+                                        ).replace(tzinfo=pytz.utc)
+        r_end = datetime.strptime(
+                                      r_end,
+                                      '%m/%d/%Y %I:%M %p'
+                                      ).replace(tzinfo=pytz.utc)
+        case_1 = Reservation.objects.filter(
+                                            table=table,
+                                            reserve_start__lte=r_start,
+                                            reserve_end__gte=r_start
+                                            ).exists()
+        case_2 = Reservation.objects.filter(
+                                            table=table,
+                                            reserve_start__lte=r_end,
+                                            reserve_end__gte=r_end
+                                            ).exists()
+        case_3 = Reservation.objects.filter(
+                                            table=table,
+                                            reserve_start__gte=r_start,
+                                            reserve_end__lte=r_end
+                                            ).exists() 
         if case_1 or case_2 or case_3:
-            messages.warning(request, 'This table is not available on your booking schedule.')
+            messages.warning(
+                request,
+                'This table is not available on your booking schedule.'
+                )
             return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
-        elif now() > reserve_start or now() > reserve_end or reserve_start > reserve_end:
-            messages.error(request, 'You can not book from past,please check your inputs.')
-            return HttpResponseRedirect(request.META.get('HTTP_REFERER'))         
+        elif now() > r_start or now() > r_end or r_start > r_end:
+            messages.error(
+                request,
+                'You can not book from past,please check your inputs.'
+                )
+            return HttpResponseRedirect(request.META.get('HTTP_REFERER')) 
         else:
             reservation = Reservation.objects.create(
                customer=customer,
                table=table,
-               reserve_start=reserve_start,
-               reserve_end=reserve_end
+               reserve_start=r_start,
+               reserve_end=r_end
             )
             reservation.save()
-            messages.success(request, 'You have successfully booked the table.')
+            messages.success(
+                request,
+                'You have successfully booked the table.'
+                )
             return redirect('restaurant:home')
     return render(request, 'restaurant/table_detail.html', context)
 
@@ -77,7 +112,7 @@ def reservation_list(request):
     user = request.user
     profile = get_object_or_404(Profile, user=user)
     reservations = Reservation.objects.filter(customer=user)
-    context = {'profile': profile, 'reservations': reservations }
+    context = {'profile': profile, 'reservations': reservations}
     return render(request, 'restaurant/reservation_list.html', context)
 
 
@@ -88,7 +123,6 @@ def reservation_edit(request, id):
         of a logged in user.
     '''
     r = get_object_or_404(Reservation, id=id)
-    # table_images = TableImage.objects.filter(table=r.table)
     form = ReservationForm(instance=r)
     context = {'r': r, 'form': form}
     if now() > r.reserve_end:
@@ -98,10 +132,16 @@ def reservation_edit(request, id):
         form = ReservationForm(request.POST, instance=r)
         if form.is_valid():
             form.save()
-            messages.success(request, 'You have successfully updated your reservation.')
+            messages.success(
+                request,
+                'You have successfully updated your reservation.'
+                )
             return redirect('restaurant:reservations')
         else:
-            messages.error(request, 'Please enter valid values to input fields.')
+            messages.error(
+                request,
+                'Please enter valid values to input fields.'
+                )
             return redirect('restaurant:reservations')
     return render(request, 'restaurant/reservation_edit.html', context)
 
@@ -109,13 +149,18 @@ def reservation_edit(request, id):
 @login_required(login_url='account/signin')
 def reservation_delete(request, id):
     '''
-        This view enables a logged in user to delete 
-        his/her reservation.
+        This view displays the confirmation page
+        if a user wishes to delete.
     '''
     r = get_object_or_404(Reservation, id=id)
-    r.delete()
-    messages.success(request, f'You have successfully deleted your reservation of {r.table.title }.')
-    return redirect('account:profile')
+    if request.method == 'POST':
+        r.delete()
+        messages.success(
+                    request,
+                    f'You have successfully deleted your reservation of {r.table.title }.'
+                    )
+        return redirect('account:profile')
+    return render(request, 'restaurant/confirm_delete.html')
 
 
 def review_list(request):
@@ -143,13 +188,17 @@ def review_table(request, slug):
     form = ReviewForm()
     table = get_object_or_404(Table, slug=slug)
     reviews = Review.objects.filter(table=table)
-    context = {'table': table,'form': form, 'reviews':reviews}
+    context = {'table': table, 'form': form, 'reviews': reviews}
     if request.method == 'POST':
         form = ReviewForm(request.POST)
         if form.is_valid():
             content = request.POST.get('content')
             author = request.user
-            review = Review.objects.create(table=table, author=author, content=content)
+            review = Review.objects.create(
+                                         table=table,
+                                         author=author,
+                                         content=content
+                                         )
             review.save()
             messages.success(request, 'Your review is added successfully.')
             return redirect('restaurant:home')
@@ -161,7 +210,7 @@ def review_table(request, slug):
 
 def login_user(request):
     '''
-        This view enables users to sign in 
+        This view enables users to sign in
         to view pages allowed only for logged in users.
     '''
     return render(request, 'restaurant/login_user.html')
@@ -173,7 +222,6 @@ def food_list(request):
     '''
     desserts = Food.objects.filter(food_type='dessert')
     foods = Food.objects.filter(food_type='main')
-    
     context = {'desserts': desserts, 'foods': foods}
     return render(request, 'restaurant/food_list.html', context)
